moved action environment activation handler
